# docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateDescribeTaskReq:
    properties:
      callback_url:
        type: string
      image_url:
        type: string
    required:
    - image_url
    type: object
  handler.CreateImagineTaskReq:
    properties:
      callback_url:
        type: string
      prompt:
        type: string
    required:
    - prompt
    type: object
  handler.PerformTaskActionReq:
    properties:
      action_type:
        description: Operation type
        type: string
      callback_url:
        type: string
      index:
        description: 'Index: 1-4'
        type: integer
      task_id:
        description: Original task ID
        type: string
    required:
    - action_type
    - index
    - task_id
    type: object
  handler.TaskListResponse:
    properties:
      tasks:
        items:
          $ref: '#/definitions/handler.TaskView'
        type: array
      total:
        type: integer
    type: object
  handler.TaskView:
    properties:
      created_at:
        type: string
      description:
        type: string
      error_message:
        type: string
      finished_at:
        type: string
      image_url:
        type: string
      oss_image_url:
        type: string
      parent_task_id:
        type: string
      progress:
        type: integer
      prompt:
        type: string
      status:
        $ref: '#/definitions/model.TaskStatus'
      task_id:
        type: string
      type:
        $ref: '#/definitions/model.TaskType'
      updated_at:
        type: string
    type: object
...
  model.TaskStatus:
    enum:
    - PENDING
    - SUBMITTED
    - IN_QUEUE
    - PROCESSING
    - SUCCESS
    - FAILED
    - TIMEOUT
    type: string
    x-enum-varnames:
    - TaskStatusPending
    - TaskStatusSubmitted
    - TaskStatusInQueue
    - TaskStatusProcessing
    - TaskStatusSuccess
    - TaskStatusFailed
    - TaskStatusTimeout
  model.TaskType:
    enum:
    - IMAGINE
    - DESCRIBE
    - UPSCALE
    - ZOOM_OUT_2X
    - ZOOM_OUT_1_5X
    - UPSCALE_SUBTLE
    - UPSCALE_CREATIVE
    type: string
    x-enum-comments:
      TaskTypeDescribe: Describe image
      TaskTypeUpscale: Upscale
      TaskTypeUpscaleCreative: Upscale (Creative)
      TaskTypeUpscaleSubtle: Upscale (Subtle)
      TaskTypeZoomOut1_5x: Zoom Out 1.5x
      TaskTypeZoomOut2x: Zoom Out 2x
    x-enum-descriptions:
    - ""
    - Describe image
    - Upscale
    - Zoom Out 2x
    - Zoom Out 1.5x
    - Upscale (Subtle)
    - Upscale (Creative)
    x-enum-varnames:
    - TaskTypeImagine
    - TaskTypeDescribe
    - TaskTypeUpscale
    - TaskTypeZoomOut2x
    - TaskTypeZoomOut1_5x
    - TaskTypeUpscaleSubtle
    - TaskTypeUpscaleCreative
  response.Response:
    properties:
      code:
        type: string
      data: {}
      detail:
        type: string
      message:
        type: string
    type: object
  service.ProcessingQueueItem:
    properties:
      created_at:
        type: string
      description:
        type: string
      finished_at:
        type: string
      image_url:
        type: string
      oss_image_url:
        type: string
      parent_task_id:
        type: string
      progress:
        type: integer
      prompt:
        type: string
      status:
        $ref: '#/definitions/model.TaskStatus'
      task_id:
        type: string
      type:
        $ref: '#/definitions/model.TaskType'
      updated_at:
        type: string
    type: object
  service.QueueItem:
    properties:
      image_url:
        type: string
      kind:
        $ref: '#/definitions/service.QueueMessageKind'
      parent_task_id:
        type: string
      prompt:
        type: string
      task_id:
        type: string
    type: object
  service.QueueMessageKind:
    enum:
    - imagine
    - describe
    - action
    type: string
    x-enum-varnames:
    - QueueMessageKindImagine
    - QueueMessageKindDescribe
    - QueueMessageKindAction
  service.QueueStatus:
    properties:
      processing_count:
        type: integer
      processing_tasks:
        items:
          $ref: '#/definitions/service.ProcessingQueueItem'
        type: array
      queue_length:
        type: integer
      waiting_in_queue:
        items:
          $ref: '#/definitions/service.QueueItem'
        type: array
    type: object
  service.TaskResponse:
    properties:
      status:
        $ref: '#/definitions/model.TaskStatus'
      task_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Midjourney Proxy API Service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Midjourney Proxy API
  version: "1.0"
paths:
...
  /api/v1/tasks:
    get:
      description: Get the task list
      parameters:
      - default: 10
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.TaskListResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: List tasks
      tags:
      - Task
  /api/v1/tasks/{task_id}:
    get:
      description: Get task details by task ID
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.TaskView'
              type: object
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get task status
      tags:
      - Task
  /api/v1/tasks/action:
    post:
      consumes:
      - application/json
      description: Perform an action on a completed task (upscale, zoom_out_2x, zoom_out_1_5x,
        upscale_subtle, upscale_creative)
      parameters:
      - description: Action request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.PerformTaskActionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Action submitted
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.TaskResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Perform task action
      tags:
      - Task
  /api/v1/tasks/describe:
    post:
      consumes:
      - application/json
      description: Create a Midjourney describe task that generates text prompts based
        on an image URL
      parameters:
      - description: Describe task request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateDescribeTaskReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.TaskResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create Describe task
      tags:
      - Task
  /api/v1/tasks/imagine:
    post:
      consumes:
      - application/json
      description: Create a Midjourney image generation task based on the given prompt
      parameters:
      - description: Task request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateImagineTaskReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.TaskResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create Imagine task
      tags:
      - Task
  /api/v1/tasks/queue:
    get:
      description: Get the list of tasks currently waiting to be processed
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.QueueStatus'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get queue list
      tags:
      - Task
...
schemes:
- http
- https
swagger: "2.0"